+++ src/tests/macros_test.rs
@@ -56,10 +56,16 @@ fn test_init_fli_from_toml_can_add_commands() {
 
 #[test]
 fn test_init_fli_from_toml_can_add_options() {
-    use crate::option_parser::ValueTypes;
+    use crate::option_parser::{Value, ValueTypes};
 
     let mut app = crate::init_fli_from_toml!();
-    app.add_option("test", "Test option", "-t", "--test", ValueTypes::None);
+    app.add_option(
+        "test",
+        "Test option",
+        "-t",
+        "--test",
+        ValueTypes::OptionalSingle(Some(Value::Bool(false))),
+    );
 
     assert!(app.root_command.get_option_parser().has_option("-t"));
 }